Added post album, Clickoutside
A user is now able to click on a post to open it in full screen and use the arrow keys beside the post to got to previous or next post. On clicking outside a div like notifications, it will now close. Added border radius to make look less aggressive.